Release 1.1.0 - Issue with bid (float value)

+++ action/bid.php
@@ -21,7 +21,7 @@ if (isset($_SESSION['mail']) AND isset($_GET['id']) AND isset($_POST['prix'])) {
     if ($req->rowCount() >= 1) { // L'objet existe
         $res = $req->fetch(PDO::FETCH_BOTH);
 
-        if ($res[1] < intval($_POST['prix'])) {
+        if ($res[1] < floatval($_POST['prix'])) {
             if ($_SESSION['bid_count'] < 4) {
                 if ($_SESSION['_id'] != $res[2]) {
                     try {
